% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_report}
\alias{add_report}
\title{Add a report obtained from a data cleaning step to a data frame}
\usage{
add_report(data, report, name = NULL)
}
\arguments{
\item{data}{A  data frame}

\item{report}{A named list or a data frame containing details from the
cleaning operations.}

\item{name}{A character with the name of the cleaning operation. The default
value is NULL.}
}
\value{
The input data frame with a report associated to it. This can be
accessed using \code{attr(data, "report")}
}
\description{
Add a report obtained from a data cleaning step to a data frame
}
\keyword{internal}
