% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary_based_cleaning.R
\name{add_to_dictionnary}
\alias{add_to_dictionnary}
\title{Add an element to the data dictionary}
\usage{
add_to_dictionnary(dictionary, option, value, grp, order = NULL)
}
\arguments{
\item{dictionary}{a data frame with the data dictionary}

\item{option}{a \code{character} with the new option to add to the dictionary}

\item{value}{a \code{character} with the value the new option should be replaced
with}

\item{grp}{a \code{character} with the name of the column that contains the option
of interest}

\item{order}{a \code{numeric} with the order of the new option}
}
\value{
an object of type data frame. This is the new data dictionary with an
additional line that contains the details about the new options.
}
\description{
Add an element to the data dictionary
}
\examples{
test <- add_to_dictionnary(
  dictionary = readRDS(system.file("extdata", "test_dict.RDS",
                       package = "cleanepi")),
  option     = "ml",
  value      = "male",
  grp        = "gender",
  order      = NULL
 )
}
