% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardisation_helpers.R
\name{as_date}
\alias{as_date}
\title{Reformat the date}
\usage{
as_date(
  x,
  format = c("ymd", "ydm", "dmy", "mdy", "myd", "dym", "Ymd", "Ydm", "dmY", "mdY", "mYd",
    "dYm")
)
}
\arguments{
\item{x}{the string of interest}

\item{format}{the date format}
}
\description{
Reformat the date
}
