% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_column_existence}
\alias{check_column_existence}
\title{Check if date column exists in given data frame}
\usage{
check_column_existence(data, date_column_name)
}
\arguments{
\item{data}{the input data frame}

\item{date_column_name}{the date column of interest}
}
\description{
Check if date column exists in given data frame
}
