% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_first_and_last_date.R
\name{check_first_and_last_date}
\alias{check_first_and_last_date}
\title{Check date time frame}
\usage{
check_first_and_last_date(first_date, last_date)
}
\arguments{
\item{first_date}{a Date object specifying the first valid date. Defaults to
fifty years before the \code{last_date}}

\item{last_date}{last_date a Date object specifying the last valid date.
Defaults to the current date.}
}
\value{
a list with the first and last date
}
\description{
Check date time frame
}
