% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_subject_ids.R
\name{check_id_length}
\alias{check_id_length}
\title{Check the length of sample IDs}
\usage{
check_id_length(x, ref)
}
\arguments{
\item{x}{the sample ID}

\item{ref}{the expected length of the sample ids}
}
\description{
Check the length of sample IDs
}
\keyword{internal}
