% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_subject_ids.R
\name{check_ids_uniqueness}
\alias{check_ids_uniqueness}
\title{Checks the uniqueness in values of the sample IDs column}
\usage{
check_ids_uniqueness(data, id_col_name)
}
\arguments{
\item{data}{the input data frame}

\item{id_col_name}{the name of the column that contains the sample IDs}
}
\value{
the input data frame with and extra element in its attributes when
there are missing or duplicated IDs.
}
\description{
Checks the uniqueness in values of the sample IDs column
}
\keyword{internal}
