% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_prefix.R
\name{check_prefix}
\alias{check_prefix}
\title{Check the prefix of the subject IDs}
\usage{
check_prefix(x, prefix)
}
\arguments{
\item{x}{the sample ID}

\item{prefix}{the prefix to look up to}
}
\description{
Check the prefix of the subject IDs
}
