% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_subject_ids.R
\name{check_subject_ids}
\alias{check_subject_ids}
\title{Check whether the subject IDs comply with the expected format}
\usage{
check_subject_ids(
  data,
  format,
  id_column_name,
  prefix = NULL,
  suffix = NULL,
  range = NULL
)
}
\arguments{
\item{data}{the data frame of interest}

\item{format}{the expected format of the subject IDs}

\item{id_column_name}{the name of the column with the subject IDs. If not
specified, the first column will be considered by default.}

\item{prefix}{the prefix used in the subject IDs}

\item{suffix}{the suffix used in the subject IDs}

\item{range}{a vector with the range of numbers in the sample IDs}
}
\value{
the cleaned data frame with correct subject IDs. The incorrect
subject ids will be stored in the report object.
}
\description{
Check whether the subject IDs comply with the expected format
}
\examples{
dat <- check_subject_ids(
  data           = readRDS(system.file("extdata", "test_df.RDS",
                                       package = "cleanepi")),
  id_column_name = "study_id",
  format         = "PS000P2",
  prefix         = "PS",
  suffix         = "P2",
  range          = c(1, 100)
)
}
