% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_subject_ids.R
\name{check_subject_ids_length}
\alias{check_subject_ids_length}
\title{Check the length of sample IDs}
\usage{
check_subject_ids_length(x, ref)
}
\arguments{
\item{x}{the sample ID}

\item{ref}{the template sample ID}
}
\description{
Check the length of sample IDs
}
\keyword{internal}
