% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_subject_ids.R
\name{check_subject_ids_oness}
\alias{check_subject_ids_oness}
\title{Checks the uniqueness in values of the sample IDs column}
\usage{
check_subject_ids_oness(data, id_col_name)
}
\arguments{
\item{data}{The input \code{<data.frame>} or \code{<linelist>}}

\item{id_col_name}{A \code{<character>} with the name of the column that
contains the sample IDs}
}
\value{
the input \code{<data.frame>} with and extra element in its
attributes when there are missing or duplicated IDs.
}
\description{
Checks the uniqueness in values of the sample IDs column
}
\keyword{internal}
