% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_subject_ids.R
\name{check_suffix}
\alias{check_suffix}
\title{Check the suffix of the subject IDs}
\usage{
check_suffix(x, suffix)
}
\arguments{
\item{x}{the sample IDs}

\item{suffix}{the suffix to lookup to}
}
\description{
Check the suffix of the subject IDs
}
