% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_stuffs.R
\name{clean_subject_ids}
\alias{clean_subject_ids}
\title{function to clean the subject ID column in a data frame}
\usage{
clean_subject_ids(
  data = NULL,
  id.position = 1,
  format = NULL,
  check = TRUE,
  prefix = NULL,
  suffix = NULL,
  range = NULL
)
}
\arguments{
\item{data}{the data frame of interest}

\item{id.position}{the column position of the variable that unique identifies the subjects. This should only be specified when the column with the subject IDs is not the first column. default is 1.}

\item{format}{the correct format for the subject IDs}

\item{check}{a boolean that specifies whether to clean the subject IDs or not}

\item{prefix}{the prefix used in the subject IDs}

\item{suffix}{the prefix used in the subject IDs}

\item{range}{a vector with the range of numbers in the sample IDs}
}
\value{
the input data frame with the corrected subject IDs
}
\description{
function to clean the subject ID column in a data frame
}
