% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_stuffs.R
\name{cleanepi}
\alias{cleanepi}
\title{function to clean epidemiological data}
\usage{
cleanepi(
  data,
  clean.col.names = FALSE,
  standardise.date = FALSE,
  date.columns = NULL,
  format = NULL,
  remove.duplicates = FALSE,
  duplicates.from = NULL,
  remove.empty = FALSE,
  remove.constant = FALSE,
  replace.missing = FALSE,
  na.comes.as = NULL,
  replace.na.at = NULL,
  replace.na.in.type = NULL
)
}
\description{
function to clean epidemiological data
}
