% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_numeric_to_date.R
\name{convert_numeric_to_date}
\alias{convert_numeric_to_date}
\title{convert numeric into Date}
\usage{
convert_numeric_to_date(data, numeric_date_columns, ref_date, forward = TRUE)
}
\arguments{
\item{data}{the input data frame}

\item{numeric_date_columns}{a vector or a comma-separated list of the columns
names to be converted from numeric to date}

\item{ref_date}{the reference date}

\item{forward}{whether the counts started after the reference date (TRUE) or
not (FALSE). default is FALSE}
}
\value{
the input data frame where the column of interest are updated
}
\description{
convert numeric into Date
}
