% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_numeric.R
\name{convert_to_numeric}
\alias{convert_to_numeric}
\title{Convert columns into numeric}
\usage{
convert_to_numeric(data, to_numeric = NULL, scan_res = NULL)
}
\arguments{
\item{data}{A data frame}

\item{to_numeric}{A vector of column names}

\item{scan_res}{A data frame obtained from the \code{scan_data()} function}
}
\value{
A resulting data frame after the conversion process,
wherein all the specified or detected columns have been transformed into
numeric format.
}
\description{
When the function is invoked without specifying the column names to be
converted, it automatically scans for columns containing exclusively missing,
numeric, and character values. Furthermore, it identifies columns where the
proportion of numeric values is at least twice the percentage of character
values.
}
\examples{
dat <- convert_to_numeric(
  data = readRDS(system.file("extdata", "messy_data.RDS",
                             package = "cleanepi")),
  to_numeric = "age",
  scan_res   = NULL
)
}
