% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardization_helpers.R
\name{date_check_column_existence}
\alias{date_check_column_existence}
\title{Check if date column exists in the given dataset}
\usage{
date_check_column_existence(data, date_column_names)
}
\arguments{
\item{data}{The input data frame}

\item{date_column_names}{A vector with the name of the columns to check}
}
\value{
The input vector if all column names are part of the input data, an
error is issued otherwise.
}
\description{
Check if date column exists in the given dataset
}
\keyword{internal}
