% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardization_helpers.R
\name{date_check_outsiders}
\alias{date_check_outsiders}
\title{Convert and update the date values}
\usage{
date_check_outsiders(data, timeframe, new_dates, cols)
}
\arguments{
\item{data}{A data frame or linelist}

\item{timeframe}{A vector of 2 values of type date. If provided, date values
that do not fall within this timeframe will be set to \code{NA}.}

\item{new_dates}{A vector of the converted date values}

\item{cols}{The names of the date column been converted}
}
\value{
A list of 2 data frames: the updated input data (if some columns
were converted to Date) and a data frame of date values that are not
within the specified timeframe.
}
\description{
Convert and update the date values
}
\keyword{internal}
