% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardisation_helpers.R
\name{date_check_timeframe}
\alias{date_check_timeframe}
\title{Check date time frame}
\usage{
date_check_timeframe(first_date, last_date)
}
\arguments{
\item{first_date}{a Date object specifying the first valid date. Defaults to
fifty years before the \code{last_date}}

\item{last_date}{last_date a Date object specifying the last valid date.
Defaults to the current date.}
}
\value{
a list with the first and last date
}
\description{
Check date time frame
}
\keyword{internal}
