% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardization_helpers.R
\name{date_detect_day_or_month}
\alias{date_detect_day_or_month}
\title{Detect the appropriate abbreviation for day or month value}
\usage{
date_detect_day_or_month(x)
}
\arguments{
\item{x}{The input string}
}
\value{
A string with abbreviation used to distinguish the written day or
month
}
\description{
Detect the appropriate abbreviation for day or month value
}
\keyword{internal}
