% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardisation_helpers.R
\name{date_detect_simple_format}
\alias{date_detect_simple_format}
\title{Get format from a simple date value}
\usage{
date_detect_simple_format(x)
}
\arguments{
\item{x}{A string with the date value of interest}
}
\description{
Get format from a simple date value
}
\keyword{internal}
