% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardisation_helpers.R
\name{date_get_format}
\alias{date_get_format}
\title{Detect date format from a date column}
\usage{
date_get_format(data, date_column_name, sep)
}
\arguments{
\item{data}{the input data frame}

\item{date_column_name}{the name of the date columns of interest}

\item{sep}{the separator in the date string}
}
\value{
a string with the detected date format
}
\description{
Detect date format from a date column
}
\keyword{internal}
