% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_get_part1}
\alias{date_get_part1}
\title{Split a string based on a pattern and return the first element of the
resulting vector.}
\usage{
date_get_part1(x, sep)
}
\arguments{
\item{x}{A string of interest}

\item{sep}{A separator in the string of interest}
}
\value{
A string with the first element of the vector returned by the
\code{strsplit()} function.
}
\description{
Split a string based on a pattern and return the first element of the
resulting vector.
}
\keyword{internal}
