% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_get_part2}
\alias{date_get_part2}
\title{Get part2 of date value}
\usage{
date_get_part2(x, sep)
}
\arguments{
\item{x}{A \code{<character>} or \code{<Date>} value}

\item{sep}{A \code{<character>} with the separator identified in 'x'}
}
\value{
A \code{<character>} with the second element of the vector returned
by the \code{strsplit()} function.
}
\description{
Get part2 of date value
}
\keyword{internal}
