% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_get_part3}
\alias{date_get_part3}
\title{Get part3 of date value}
\usage{
date_get_part3(x, sep)
}
\arguments{
\item{x}{\code{<character>} or \code{<Date>} value}

\item{sep}{A \code{<character>} with the separator identified in 'x'}
}
\value{
A \code{<character>} with the third element of the vector returned
by the \code{strsplit()} function.
}
\description{
Get part3 of date value
}
\keyword{internal}
