% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardisation_helpers.R
\name{date_guess_convert}
\alias{date_guess_convert}
\title{Guess if a character vector contains Date values, and convert them into Date}
\usage{
date_guess_convert(data, error_tolerance, timeframe)
}
\arguments{
\item{data}{the input data frame}

\item{error_tolerance}{a number between 0 and 1 indicating the proportion of
entries which cannot be identified as dates to be tolerated. See the
\code{clean_data()} helper for more details}

\item{timeframe}{the expected first and last date. See the \code{clean_data()}
helper for more details}
}
\description{
Guess if a character vector contains Date values, and convert them into Date
}
\keyword{internal}
