% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_dates.R
\name{date_i_extract_string}
\alias{date_i_extract_string}
\title{Extract date from a character string}
\usage{
date_i_extract_string(x)
}
\value{
Either \code{NA_character_} or a date, as a standardized character
string.
}
\description{
This function looks for a well-formatted date character string inside a
single character string, and returns the matching date using the \verb{\%Y-\%m-\%d}
format (e.g. \code{2018-01-23}).
}
\author{
Thibaut Jombart
}
\keyword{internal}
