% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_dates.R
\name{date_i_guess_and_convert}
\alias{date_i_guess_and_convert}
\title{Extract date from a character vector}
\usage{
date_i_guess_and_convert(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
If the format cannot be resolved, the function returns NA; if a
matching format is found, it returns the converted values.
}
\description{
This function tries converting a single character string into a
well-formatted date, but still returning a character. If it can't convert
it, it returns NA.
}
\keyword{internal}
