% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardization_helpers.R
\name{date_make_format}
\alias{date_make_format}
\title{Build the auto-detected format}
\usage{
date_make_format(f1, f2, f3)
}
\arguments{
\item{f1}{A \code{<character>} with the first part of the inferred format}

\item{f2}{A \code{<character>} with the second part of the inferred format}

\item{f3}{A \code{<character>} with the third part of the inferred format}
}
\value{
A \code{<character>} that represents the inferred format from the
provided elements. It returns \code{<NULL>} when the format was not
resolved.
}
\description{
Put together the different date format characters that were identified from
the target date column.
}
\keyword{internal}
