% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardisation_helpers.R
\name{date_match_format_and_column}
\alias{date_match_format_and_column}
\title{Check whether the number of provided formats matches the number of target
columns to be standardized.}
\usage{
date_match_format_and_column(target_columns, format)
}
\arguments{
\item{target_columns}{a vector of column names to be standardised}

\item{format}{a vector of formats to be used when standardising the columns}
}
\value{
a vector of format
}
\description{
Check whether the number of provided formats matches the number of target
columns to be standardized.
}
\keyword{internal}
