% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardization_helpers.R
\name{date_trim_outliers}
\alias{date_trim_outliers}
\title{Trim dates outside of the defined boundaries}
\usage{
date_trim_outliers(new_dates, dmin, dmax, cols, original_dates)
}
\arguments{
\item{new_dates}{A vector of the new date values}

\item{dmin}{A minimum date}

\item{dmax}{A maximum date}

\item{cols}{The name of the date column of interest}

\item{original_dates}{A vector of the original date values}
}
\value{
A list of 2 elements: the update input vector where date values that
are out of the boundaries are replaced by \code{NA}, and a vector of the out of
boundaries values.
}
\description{
Trim dates outside of the defined boundaries
}
\keyword{internal}
