% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_cleanepi_settings.R
\name{default_cleanepi_settings}
\alias{default_cleanepi_settings}
\title{Set \code{clean_data()} default parameters}
\usage{
default_cleanepi_settings()
}
\value{
The list of the default cleaning parameters.
}
\description{
When calling \code{clean_data()} function with \code{param = NULL}, the arguments
defined here will applied on the input data. By default, operations that
require the target columns to specified by the user will not be performed.
The default cleaning operations include: i) standardizing column names,
ii) detecting and removing duplicates, and iii) removing constant data.
}
\keyword{internal}
