% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_numeric.R
\name{detect_columns_to_convert}
\alias{detect_columns_to_convert}
\title{Detect the numeric columns that appears as characters due to the presence of
some character values in the column.}
\usage{
detect_columns_to_convert(scan_res)
}
\arguments{
\item{scan_res}{a data frame that corresponds to the result from the
\code{scan_data()} function}
}
\value{
a vector of column names to be converted into numeric
}
\description{
Detect the numeric columns that appears as characters due to the presence of
some character values in the column.
}
\keyword{internal}
