% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_complex_format.R
\name{detect_complex_format}
\alias{detect_complex_format}
\title{Detect complex format}
\usage{
detect_complex_format(x)
}
\arguments{
\item{x}{the string of interest}
}
\description{
Detect complex format
}
