% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_date_format.R
\name{detect_date_format}
\alias{detect_date_format}
\title{Detect the date format with only 1 separator}
\usage{
detect_date_format(x)
}
\arguments{
\item{x}{the string of interest}
}
\description{
Detect the date format with only 1 separator
}
