% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_cleanepi.R
\name{detect_date_separator}
\alias{detect_date_separator}
\title{function to detect the special character that is the separator in the date values}
\usage{
detect_date_separator(x)
}
\description{
function to detect the special character that is the separator in the date values
}
