% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_day_or_month.R
\name{detect_day_or_month}
\alias{detect_day_or_month}
\title{Detect whether it's day or month}
\usage{
detect_day_or_month(x)
}
\arguments{
\item{x}{the string of interest}
}
\description{
Detect whether it's day or month
}
