% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary_based_cleaning.R
\name{detect_misspelled_options}
\alias{detect_misspelled_options}
\title{Detect misspelled options in columns to be cleaned}
\usage{
detect_misspelled_options(data, dictionary)
}
\arguments{
\item{data}{The input data frame or linelist}

\item{dictionary}{A data dictionary associated with the input data}
}
\value{
A list with the indexes of the misspelled values in every column
that needs to be cleaned.
}
\description{
Detect misspelled options in columns to be cleaned
}
\keyword{internal}
