% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_simple_format.R
\name{detect_simple_format}
\alias{detect_simple_format}
\title{function to get simple format}
\usage{
detect_simple_format(x)
}
\arguments{
\item{x}{the string of interest}
}
\description{
function to get simple format
}
