% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_numeric.R
\name{detect_to_numeric_columns}
\alias{detect_to_numeric_columns}
\title{Detect the numeric columns that appears as characters due to the presence of
some character values in the column.}
\usage{
detect_to_numeric_columns(scan_res, data)
}
\arguments{
\item{scan_res}{A data frame that corresponds to the result from the
\code{scan_data()} function}

\item{data}{The input data from which 'scan_res' was generated.}
}
\value{
a vector of column names to be converted into numeric
}
\description{
Detect the numeric columns that appears as characters due to the presence of
some character values in the column.
}
\keyword{internal}
