% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_dates.R
\name{find_and_constrain_date}
\alias{find_and_constrain_date}
\title{Finds dates with lubridate and constrains them to a date range}
\usage{
find_and_constrain_date(
  x,
  check_timeframe,
  dmin,
  dmax,
  baddies,
  orders = NULL,
  keep = TRUE
)
}
\arguments{
\item{x}{a character vector that can be converted to dates}

\item{check_timeframe}{a logical to check whether dates fall within timeframe}

\item{dmin}{the minimum dates}

\item{dmax}{the maximum dates}

\item{baddies}{an environment that will act as a list of bad dates.}

\item{orders}{a vector of lubridate orders to consider}

\item{keep}{a logical vector indicating the dates to test from \code{x}}
}
\description{
This takes a character vector and returns a vector of successfully translated
dates within the specified range.
}
\details{
It will also have the side-effect of populating an environment of bad dates
that can be used for warning the user
}
\keyword{internal}
