% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/find_and_remove_duplicates.R
\name{find_duplicates}
\alias{find_duplicates}
\title{Identify and return duplicated rows in a data frame or linelist.}
\usage{
find_duplicates(data, target_columns)
}
\arguments{
\item{data}{The input data frame or linelist.}

\item{target_columns}{A vector of columns names or indices to consider when
=======
% Please edit documentation in R/remove_duplicates.R
\name{find_duplicates}
\alias{find_duplicates}
\title{Identify duplicated rows in a data frame or linelist}
\usage{
find_duplicates(data, duplicates_from)
}
\arguments{
\item{data}{the input data frame or linelist}

\item{duplicates_from}{a vector of columns names or indices to consider when
>>>>>>> 433caea (update NAMESPACE after devtools::document())
looking for duplicates. When the input data is a \code{linelist} object,
this parameter can be set to \code{tags} if you wish to look for
duplicates across the tagged variables only.}
}
\value{
<<<<<<< HEAD
Data frame or linelist of all duplicated rows with following 2
=======
a data frame or linelist of all duplicated rows with following 2
>>>>>>> 433caea (update NAMESPACE after devtools::document())
additional columns:
\enumerate{
\item \code{row_id}: the indices of the duplicated rows from the input data.
Users can choose from these indices, which row they consider as
redundant in each group of duplicates.
\item \code{group_id}: a unique identifier associated to each group of
duplicates.
}
}
\description{
<<<<<<< HEAD
Identify and return duplicated rows in a data frame or linelist.
}
\examples{
dups <- find_duplicates(
  data = readRDS(system.file("extdata", "test_linelist.RDS",
    package = "cleanepi")),
  target_columns = c("dt_onset", "dt_report", "sex", "outcome")
=======
Identify duplicated rows in a data frame or linelist
}
\examples{
dups <- find_duplicates(
data = readRDS(system.file("extdata", "test_linelist.RDS",
  package = "cleanepi")),
duplicates_from = "tags"
>>>>>>> 433caea (update NAMESPACE after devtools::document())
)

}
