% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_helpers.R
\name{get_appropriate_format}
\alias{get_appropriate_format}
\title{Transform scanning result format into user-chosen format}
\usage{
get_appropriate_format(counts, initial_length, format)
}
\arguments{
\item{counts}{A numeric vector with the counts of the different data types}

\item{initial_length}{A numeric with the number of rows in the dataset}

\item{format}{A character with the user-specified format}
}
\description{
Transform scanning result format into user-chosen format
}
\keyword{internal}
