% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_helpers.R
\name{get_class}
\alias{get_class}
\title{Get class of a vector}
\usage{
get_class(x)
}
\arguments{
\item{x}{The input vector}
}
\value{
A character with the class of the input vector
}
\description{
Get class of a vector
}
\keyword{internal}
