% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_helpers.R
\name{get_column_composition}
\alias{get_column_composition}
\title{Get the percentage of missing and other data type values in a vector.
The considered data types are: \code{numeric}, \code{Date}, \code{character}, \code{logical}}
\usage{
get_column_composition(x)
}
\arguments{
\item{x}{a vector of 1 or a combination of the values of the type mentioned
above}
}
\value{
a vector of 4 elements that represent respectively the percent of:
missing, numeric, Date and character values found in the input vector.
}
\description{
Get the percentage of missing and other data type values in a vector.
The considered data types are: \code{numeric}, \code{Date}, \code{character}, \code{logical}
}
\keyword{internal}
