% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardisation_helpers.R
\name{get_format}
\alias{get_format}
\title{detect date format from date column}
\usage{
get_format(data, date_column_name, sep)
}
\arguments{
\item{data}{the input data frame}

\item{date_column_name}{the date column of interest}

\item{sep}{the separator in the date string}
}
\value{
a string with the detected date format
}
\description{
detect date format from date column
}
