% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_name_standardisation.R
\name{get_new_names}
\alias{get_new_names}
\title{Get the indices of the columns to be renamed}
\usage{
get_new_names(original_names, target_columns)
}
\arguments{
\item{original_names}{A vector with the column names of the input data}

\item{target_columns}{The expression that specifies how the column names
should be renamed}
}
\value{
A named vector of the indices of the columns to be renamed
}
\description{
Get the indices of the columns to be renamed
}
\keyword{internal}
