% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_part1}
\alias{get_part1}
\title{Get part1 of date value}
\usage{
get_part1(x, sep)
}
\arguments{
\item{x}{the string of interest}

\item{sep}{the separator in the string of interest}
}
\description{
Get part1 of date value
}
