% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_part2.R
\name{get_part2}
\alias{get_part2}
\title{Get part2 of date value}
\usage{
get_part2(x, sep)
}
\arguments{
\item{x}{the string of interest}

\item{sep}{the separator in the string of interest}
}
\description{
Get part2 of date value
}
