% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_part3.R
\name{get_part3}
\alias{get_part3}
\title{Get part3 of date value}
\usage{
get_part3(x, sep)
}
\arguments{
\item{x}{the string of interest}

\item{sep}{the separator in the string of interest}
}
\description{
Get part3 of date value
}
