% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sum}
\alias{get_sum}
\title{Get sum of numbers from a string}
\usage{
get_sum(x)
}
\arguments{
\item{x}{A string of interest}
}
\value{
A numeric that correspond to the sum of every digit in the provided
string.
}
\description{
Get sum of numbers from a string
}
\keyword{internal}
