% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_target_column_names}
\alias{get_target_column_names}
\title{Get the names of the columns from which duplicates will be found}
\usage{
get_target_column_names(data, target_columns, cols)
}
\arguments{
\item{data}{The input \code{<data.frame>} or \code{<linelist>}}

\item{target_columns}{A \code{<vector>} of column names. For
\code{<linelist>} data, this can be set to \code{linelist_tags}}

\item{cols}{A \code{<vector>} of empty and constant columns}
}
\value{
A \code{<vector>} with the target column names or indexes
}
\description{
Get the names of the columns from which duplicates will be found
}
\keyword{internal}
