% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_date_sequence.R
\name{is_order}
\alias{is_order}
\title{Check order of a sequence of events}
\usage{
is_order(x)
}
\arguments{
\item{x}{A string of interest}
}
\description{
Check order of a sequence of events
}
\keyword{internal}
