% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary_based_cleaning.R
\name{make_metadata}
\alias{make_metadata}
\title{Make data dictionary for 1 field}
\usage{
make_metadata(x, field_column)
}
\arguments{
\item{x}{a \code{vector} choices from the input data}

\item{field_column}{a \code{character} with the name of the field where the
choices are made.}
}
\value{
a \verb{data frame} with the dictionary in the format that is accepted by
the {matchmaker} package.
}
\description{
Make data dictionary for 1 field
}
\keyword{internal}
