% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_name_standardization.R
\name{make_unique_column_names}
\alias{make_unique_column_names}
\title{Make column names unique when duplicated column names are found after the
transformation}
\usage{
make_unique_column_names(after, kept, before, rename)
}
\arguments{
\item{after}{A \code{<vector>} with the transformed column names}

\item{kept}{A \code{<logical>} vector where column names to keep are set to
TRUE}

\item{before}{A \code{<vector>} with the initial column names}

\item{rename}{A \code{<vector>} with the indices of the column names to be
renamed}
}
\value{
An adjusted \code{<vector>} if there were duplicated names introduced
due to the transformation
}
\description{
Make column names unique when duplicated column names are found after the
transformation
}
\keyword{internal}
