% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print.R
\name{pretty_print}
\alias{pretty_print}
\title{Print the report from data cleaning}
\usage{
pretty_print(report)
}
\arguments{
\item{report}{the report object}
}
\description{
Print the report from data cleaning
}
