% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardisation_helpers.R
\name{process_dates}
\alias{process_dates}
\title{Process date variable}
\usage{
process_dates(x, first_date, last_date, check_timeframe)
}
\arguments{
\item{x}{an object of date class}

\item{first_date}{a Date object specifying the first valid date. Defaults to
fifty years before the \code{last_date}}

\item{last_date}{last_date a Date object specifying the last valid date.
Defaults to the current date.}

\item{check_timeframe}{a logical specifying whether to check if all values
fall under the specified timeframe}
}
\value{
the modified input object
}
\description{
Process date variable
}
