% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_missing_char.R
\name{replace_missing_char}
\alias{replace_missing_char}
\title{Replace missing characters with NA}
\usage{
replace_missing_char(data, column_name, na_char = NULL)
}
\arguments{
\item{data}{the input data frame}

\item{column_name}{the column in which to look for missing character for}

\item{na_char}{the character that represents the missing values in the column
of interest}
}
\value{
a data frame where missing values are represented by NA
}
\description{
Replace missing characters with NA
}
