% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_missing_values.R
\name{replace_missing_values}
\alias{replace_missing_values}
\title{Replace missing values with \code{NA}}
\usage{
replace_missing_values(
  data,
  target_columns = NULL,
  na_strings = cleanepi::common_na_strings
)
}
\arguments{
\item{data}{the input data}

\item{target_columns}{a vector of column names. If provided, the missing
values substitution will only be performed in those columns.}

\item{na_strings}{a vector of strings that represent the missing values in
the columns of interest. Default values are \code{cleanepi::common_na_strings}.
When the missing values string in the columns of interest is not part of
this predefined vector, it can be used as the value for this argument.}
}
\value{
the input data where missing values are replaced by \code{NA}
}
\description{
Replace missing values with \code{NA}
}
\examples{
cleaned_data <- replace_missing_values(
  data        = readRDS(system.file("extdata", "test_df.RDS",
                                    package = "cleanepi")),
  target_columns = "sex",
  na_strings     = "-99"
)

}
