% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_missing_values.R
\name{replace_with_na}
\alias{replace_with_na}
\title{Detect and replace values with \code{NA} from a vector}
\usage{
replace_with_na(x, na_strings)
}
\arguments{
\item{x}{A \code{<vector>} of numeric or character values}

\item{na_strings}{A \code{<vector>} of the values to be replaced}
}
\value{
A \code{<vector>} where the specified values were replaced with
\code{NA} if found.
}
\description{
Detect and replace values with \code{NA} from a vector
}
\keyword{internal}
