% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retrieve_column_names}
\alias{retrieve_column_names}
\title{Get column names}
\usage{
retrieve_column_names(data, target_columns)
}
\arguments{
\item{data}{The input \code{<data.frame>} or \code{<linelist>}}

\item{target_columns}{A \code{<vector>} of column names. For
\code{<linelist>} data, this can be set to \code{linelist_tags}}
}
\value{
A \code{<vector>} of column names to be used for the target cleaning
operations
}
\description{
When performing several data cleaning operations using the
\code{clean_data()} function, the input column names might be altered by
after the column names cleaning. As a consequence of this, some cleaning
operations will fail due to the column names mismatch. This function is
provided to anticipate on this scenario, hence providing continuity between
the cleaning operations.
}
\keyword{internal}
