% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_helpers.R
\name{scan_columns}
\alias{scan_columns}
\title{Calculate the percentage of missing and other data type values in a vector
containing different data types such as numeric, Date, character, and
logical.}
\usage{
scan_columns(x)
}
\arguments{
\item{x}{A vector of ones or a combination of various data types.}
}
\value{
A vector of 5 elements representing the percentage of missing,
numeric, date, character, and logical values found in the input vector.
}
\description{
Calculate the percentage of missing and other data type values in a vector
containing different data types such as numeric, Date, character, and
logical.
}
\keyword{internal}
