% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_helpers.R
\name{scan_in_character}
\alias{scan_in_character}
\title{Scan through a character column}
\usage{
scan_in_character(x)
}
\arguments{
\item{x}{The input vector}
}
\value{
A numeric vector with the proportion of the different types of data
that were detected within the input vector.
}
\description{
Scan through a character column
}
\keyword{internal}
