% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_stuffs.R
\name{standardize_date}
\alias{standardize_date}
\title{function to standardize date variable}
\usage{
standardize_date(data, date.column.name = NULL, format = NULL)
}
\arguments{
\item{data}{a data frame with the date column to standardise}

\item{date.column.name}{the name of the date column of interest. default: 'Date', or 'DATE', or 'date'}

\item{format}{the format of the date values in the date column}
}
\value{
a data frame where the specified date column is converted into the format 'yyyy-mm-dd'
}
\description{
function to standardize date variable
}
