% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_numeric.R
\name{to_numeric_convert}
\alias{to_numeric_convert}
\title{Convert values in a character vector into numeric}
\usage{
to_numeric_convert(x)
}
\arguments{
\item{x}{a vector of type character or numeric}
}
\value{
a vector of type numeric with the same length as the input vector
}
\description{
The conversion is only applied on non-missing and non-numeric values found
from the input vector
}
\keyword{internal}
