% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_report}
\alias{add_report}
\title{Add a report obtained from a data cleaning step to a data frame}
\usage{
add_report(data, report, name = NULL)
}
\arguments{
\item{data}{the input data frame}

\item{report}{a named \code{list} or \verb{data frame} containing details from the
cleaning step}

\item{name}{a \code{character} with the name of the cleaning step. The default
value is NULL.}
}
\value{
the input data frame with a report associated to it. This can be
accessed using \code{attr(data, "report")}
}
\description{
Add a report obtained from a data cleaning step to a data frame
}
\keyword{internal}
