% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_to_report}
\alias{add_to_report}
\title{Add an element to the report object}
\usage{
add_to_report(x, key, value = NULL)
}
\arguments{
\item{x}{an object of class \verb{data frame} or \code{linelist}}

\item{key}{the name of the cleaning operation}

\item{value}{the object to add to the report object}
}
\value{
the input report object with an additional element
}
\description{
Add an element to the report object
}
\keyword{internal}
