% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcualte_age.R
\name{calculate_age}
\alias{calculate_age}
\title{Calculate age from a specified date column}
\usage{
calculate_age(
  data,
  target_column = NULL,
  end_date = Sys.Date(),
  age_in = "years"
)
}
\arguments{
\item{data}{the input data frame with the date column of interest}

\item{target_column}{the name of the date column of interest. default are
'Date', or, 'DATE', or 'date'}

\item{end_date}{the end date. default: today's date}

\item{age_in}{a string that specifies whether to return the age in 'years',
or 'months', or 'days', or 'weeks'. Default is 'years'.}
}
\value{
the input data frame with the following 1 or 2 extra columns:
\enumerate{
\item "age_in_years", or "age_in_months", or "age_in_weeks", or
"age_in_days", depending on the value of the 'age_in' parameter.
\item "remainder_days": a column with the number of remaining days after
the age is converted in weeks or months.
}
}
\description{
Calculate age from a specified date column
}
\examples{
age <- calculate_age(
  data          = readRDS(system.file("extdata", "test_df.RDS",
                                      package = "cleanepi")),
  target_column = "dateOfBirth",
  end_date      = Sys.Date(),
  age_in        = "months"
)
}
