% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcualte_age.R
\name{calculate_age}
\alias{calculate_age}
\title{function to calculate age from date of birth}
\usage{
calculate_age(
  data,
  date_column_name = NULL,
  end_date = Sys.Date(),
  age_in = "years"
)
}
\arguments{
\item{data}{a data frame with the date column to standardise}

\item{date_column_name}{the name of the date column of interest.
default: 'Date', or, 'DATE', or 'date'}

\item{end_date}{the end date. default: today's date}

\item{age_in}{a character string to specify whether to calculate
the age in 'years', or 'months', or 'days', or 'weeks'.
default is: 'years'}
}
\value{
a data frame with 1 or 2 extra columns compared to the input data
frame
\enumerate{
\item "age_years", or "age_months", or "age_weeks", or "age_days",
depending on the value of the 'age_in' parameter
\item "remainder_days": a column with the number of remaining days
after the age is converted in weeks or months
}
}
\description{
function to calculate age from date of birth
}
\examples{
age <- calculate_age(
data = readRDS(system.file("extdata", "test_df.rds", package = "cleanepi")),
date_column_name = "dateOfBirth",
end_date = Sys.Date(),
age_in = "months"
)
}
