% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_date_sequence.R
\name{check_date_sequence}
\alias{check_date_sequence}
\title{Check the sequence of event dates}
\usage{
check_date_sequence(data, event.cols, remove.bad.seq = TRUE)
}
\arguments{
\item{data}{the input data frame}

\item{event.cols}{a vector or a comma-separated list of the names of the event columns.
Users should specify at least 2 column names. These column names should be provided
with the expected order. For example: event.cols = c("DS","DH","DD") where
DS=date of the symptoms onset, DH=date of hospitalization, DD=date of death}

\item{remove.bad.seq}{a Boolean to specify if rows with incorrect order should be filtered out or not. default is TRUE}
}
\value{
a data frame where rows with incorrect date sequence have been removed if \code{remove.bad.seq=TRUE}, the input data frame if not
}
\description{
This function should be used to check whether date sequence in the columns specified by the user is correct or not.
}
