% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_id_length.R
\name{check_id_length}
\alias{check_id_length}
\title{Check the length of sample IDs}
\usage{
check_id_length(x, ref)
}
\arguments{
\item{x}{the sample ID}

\item{ref}{the template sample ID}
}
\description{
Check the length of sample IDs
}
