% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_helpers.R
\name{check_ids_uniqueness}
\alias{check_ids_uniqueness}
\title{Check the uniqueness in values of the sample IDs column}
\usage{
check_ids_uniqueness(data, id_col_name, report = list())
}
\arguments{
\item{data}{the input data frame}

\item{id_col_name}{the name of the column that contains the sample IDs}

\item{report}{the object that will contains details about the result from the
date columns standardization.}
}
\value{
a report object of type \code{list}. This should contain an element named
as \strong{duplicated_ids}. This is either a \code{data.frame} if there are rows
with duplicated IDs or \code{NULL} if not.
}
\description{
Check the uniqueness in values of the sample IDs column
}
\examples{
<<<<<<< HEAD
\dontrun{
=======
>>>>>>> 7ed29d1 (add scan_data(), and check_ids_uniqueness() functions)
report <- check_ids_uniqueness(
data = readRDS(system.file("extdata", "messy_data.RDS",
    package = "cleanepi")),
id_col_name = "case_id",
)
}
<<<<<<< HEAD
}
=======
>>>>>>> 7ed29d1 (add scan_data(), and check_ids_uniqueness() functions)
