% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_subject_ids.R
\name{check_subject_ids}
\alias{check_subject_ids}
\title{Check whether the subject IDs comply with the expected format}
\usage{
check_subject_ids(
  data,
  id_column_name = NULL,
  format,
  prefix = NULL,
  suffix = NULL,
  range = NULL,
  remove = FALSE,
  verbose = FALSE,
  report = list()
)
}
\arguments{
\item{data}{the data frame of interest}

\item{id_column_name}{the name of the column with the subject IDs. If not
specified, the first column will be considered by default}

\item{format}{the expected subject IDs format}

\item{prefix}{the prefix used in the subject IDs}

\item{suffix}{the prefix used in the subject IDs}

\item{range}{a vector with the range of numbers in the sample IDs}

\item{remove}{a Boolean to specify whether to remove rows with incorrect}

\item{verbose}{a Boolean to specify whether print the detected incorrect
subject IDs}

\item{report}{the report object}
}
\value{
if found, the function return a list with 2 elements: the cleaned
data frame with correct subject IDs and a report containing the rows of
the input data frame with incorrect subject IDs
}
\description{
Check whether the subject IDs comply with the expected format
}
\examples{
dat <- check_subject_ids(
data = readRDS(system.file("extdata", "test_df.RDS", package = "cleanepi")),
id_column_name = "study_id",
format = "PS000P2",
prefix = "PS",
suffix = "P2",
range = c(1,100),
remove = FALSE,
verbose = TRUE,
report = list()
)
}
