% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_col_names.R
\name{clean_col_names}
\alias{clean_col_names}
\title{clean column names of a data frame}
\usage{
clean_col_names(x)
}
\arguments{
\item{x}{the input data frame}
}
\value{
the input data frame with a knit column names
}
\description{
clean column names of a data frame
}
