% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary_based_cleaning.R
\name{clean_using_dictionary}
\alias{clean_using_dictionary}
\title{Perform dictionary-based cleaning}
\usage{
clean_using_dictionary(data, dictionary, correct = FALSE)
}
\arguments{
\item{data}{the input data}

\item{dictionary}{a \verb{data frame} with the data dictionary associated to the
input data. This should be in the format}

\item{correct}{a \code{logical} specifying whether to correct the misspelled
values in the data or not. default is \code{FALSE}}
}
\value{
a \verb{data frame} with cleaned values in the target columns specified
in the data dictionary.
}
\description{
Perform dictionary-based cleaning
}
\examples{
cleaned_df <- clean_using_dictionary(
  data       = readRDS(system.file("extdata", "messy_data.RDS",
                       package = "cleanepi")),
  dictionary = readRDS(system.file("extdata", "test_dict.RDS",
                       package = "cleanepi")),
  correct    = FALSE
)
}
