% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanepi-package.R
\docType{package}
\name{cleanepi-package}
\alias{cleanepi}
\alias{cleanepi-package}
\title{cleanepi: Data Cleaning Package}
\description{
cleanepi contains functions that are used for epidemiological data cleaning.
}
\author{
\strong{Maintainer}: Karim Mané \email{lshkm1@lshtm.ac.uk}

Other contributors:
\itemize{
  \item Thibaut Jombart \email{thibautjombart@gmail.com} [copyright holder]
}

}
\keyword{internal}
