% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanepi-package.R
\docType{package}
\name{cleanepi-package}
\alias{cleanepi}
\alias{cleanepi-package}
\title{cleanepi: Clean and Standardize Epidemiological Data}
\description{
Cleaning and standardizing tabular data package, tailored specifically for curating epidemiological data. It streamlines various data cleaning tasks that are typically expected when working with datasets in epidemiology. It returns the processed data in the same format, ensuring seamless integration into existing workflows. Additionally, it generates a comprehensive report detailing the outcomes of each cleaning task.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://epiverse-trace.github.io/cleanepi/}
  \item \url{https://github.com/epiverse-trace/cleanepi}
  \item Report bugs at \url{https://github.com/epiverse-trace/cleanepi/issues}
}

}
\author{
\strong{Maintainer}: Karim Mané \email{karim.mane@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-9892-2999}{ORCID})

Authors:
\itemize{
  \item Abdoelnaser Degoot \email{abdoelnaser-mahmood.degoot@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8788-2496}{ORCID})
  \item Bankolé Ahadzie \email{Bankole.Ahadzie@lshtm.ac.uk}
  \item Nuredin Mohammed \email{Nuredin.Mohammed@lshtm.ac.uk}
  \item Bubacarr Bah \email{Bubacarr.Bah1@lshtm.ac.uk} (\href{https://orcid.org/0000-0003-3318-6668}{ORCID})
}

Other contributors:
\itemize{
  \item Thibaut Jombart \email{thibautjombart@gmail.com} (Thibault contributed in development of date_guess().) [contributor]
  \item Hugo Gruson \email{hugo@data.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [contributor, reviewer]
  \item Pratik R. Gupte \email{pratik.gupte@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID}) [reviewer]
  \item James M. Azam \email{james.azam@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5782-7330}{ORCID}) [reviewer]
  \item Joshua W. Lambert \email{joshua.lambert@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5218-3046}{ORCID}) [reviewer]
  \item Chris Hartgerink \email{chris@data.org} (\href{https://orcid.org/0000-0003-1050-6809}{ORCID}) [reviewer]
  \item Andree Valle-Campos \email{avallecam@gmail.com} [reviewer, contributor]
  \item London School of Hygiene and Tropical Medicine, LSHTM [copyright holder]
  \item data.org [funder]
}

}
\keyword{internal}
