% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanepi-package.R
\docType{package}
\name{cleanepi-package}
\alias{cleanepi}
\alias{cleanepi-package}
\title{cleanepi: Data Cleaning Package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

cleanepi contains functions that are used for epidemiological data cleaning.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://epiverse-trace.github.io/cleanepi}
  \item \url{https://github.com/epiverse-trace/cleanepi}
  \item Report bugs at \url{https://github.com/epiverse-trace/cleanepi/issues}
}

}
\author{
\strong{Maintainer}: Karim Mané \email{lshkm1@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-9892-2999}{ORCID})

Other contributors:
\itemize{
  \item Thibaut Jombart \email{thibautjombart@gmail.com} [copyright holder]
  \item Bubacarr Bah \email{Bubacarr.Bah1@lshtm.ac.uk} (\href{https://orcid.org/0000-0003-3318-6668}{ORCID}) [contributor]
  \item Bankolé Ahadzie \email{Bankole.Ahadzie@lshtm.ac.uk} [contributor]
  \item Nuredin Mohammed \email{Nuredin.Mohammed@lshtm.ac.uk} [contributor]
  \item Abdoelnaser Degoot \email{abdoelnaser-mahmood.degoot@lshtm.ac.uk} [contributor]
  \item Hugo Gruson \email{hugo@data.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [reviewer]
}

}
\keyword{internal}
