% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanepi-package.R, R/package.R
\docType{package}
\name{cleanepi-package}
\alias{cleanepi-package}
\title{cleanepi: Clean and Standardize Epidemiological Data}
\description{
cleanepi provides functions for cleaning and standardizing tabular data, tailored specifically for curating epidemiological data.

cleanepi provides functions for cleaning and standardizing tabular data, tailored specifically for curating epidemiological data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://epiverse-trace.github.io/cleanepi}
  \item \url{https://github.com/epiverse-trace/cleanepi}
  \item Report bugs at \url{https://github.com/epiverse-trace/cleanepi/issues}
}


Useful links:
\itemize{
  \item \url{https://epiverse-trace.github.io/cleanepi}
  \item \url{https://github.com/epiverse-trace/cleanepi}
  \item Report bugs at \url{https://github.com/epiverse-trace/cleanepi/issues}
}

}
\author{
\strong{Maintainer}: Karim Mané \email{karim.mane@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-9892-2999}{ORCID})

Authors:
\itemize{
  \item Bubacarr Bah \email{Bubacarr.Bah1@lshtm.ac.uk} (\href{https://orcid.org/0000-0003-3318-6668}{ORCID})
  \item Bankolé Ahadzie \email{Bankole.Ahadzie@lshtm.ac.uk}
  \item Nuredin Mohammed \email{Nuredin.Mohammed@lshtm.ac.uk}
  \item Abdoelnaser Degoot \email{abdoelnaser-mahmood.degoot@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8788-2496}{ORCID})
}

Other contributors:
\itemize{
  \item Thibaut Jombart \email{thibautjombart@gmail.com} (Thibault is owner of the code in guess_dates.R file.) [copyright holder]
  \item Hugo Gruson \email{hugo@data.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [reviewer]
  \item Pratik R. Gupte \email{pratik.gupte@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID}) [reviewer]
  \item James M. Azam \email{james.azam@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5782-7330}{ORCID}) [reviewer]
  \item Joshua W. Lambert \email{joshua.lambert@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5218-3046}{ORCID}) [reviewer]
}

}
\keyword{internal}
