% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert}
\alias{convert}
\title{Convert values in a character vector into numeric}
\usage{
convert(x)
}
\arguments{
\item{x}{a vector of type character}
}
\value{
a vector of type numeric with the same length as the input vector
}
\description{
The conversion is only applied on non-missing and non-numeric values found
from the input vector
}
\examples{
\dontrun{
  x     <- c(2, "ten", NA, 2, "twenty", 6)
  num_x <- convert(x)
}

}
\keyword{internal}
