% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_numeric_to_date.R
\name{convert_numeric_to_date}
\alias{convert_numeric_to_date}
\title{Convert numeric to date}
\usage{
convert_numeric_to_date(data, numeric_date_columns, ref_date, forward = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{numeric_date_columns}{A vector or a comma-separated list of  columns
names to be converted from numeric to date}

\item{ref_date}{A reference date}

\item{forward}{A Boolean to indicate whether the counts started after the
reference date (TRUE) or not (FALSE). The default is FALSE.}
}
\value{
A data frame where the column of interest are updated
}
\description{
Convert numeric to date
}
