% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardisation_helpers.R
\name{convert_to_date}
\alias{convert_to_date}
\title{Convert characters to dates}
\usage{
convert_to_date(data, cols, sep)
}
\arguments{
\item{data}{the input data frame}

\item{cols}{the date column name(s)}

\item{sep}{the separator in the date values}
}
\value{
the input data frame where the specified columns have been converted
into Date.
}
\description{
Convert characters to dates
}
\keyword{internal}
