% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_numeric}
\alias{convert_to_numeric}
\title{Convert provided or auto-detected character columns into numeric}
\usage{
convert_to_numeric(data, report = NULL, to_numeric = NULL, scan_res = NULL)
}
\arguments{
\item{data}{the input data frame}

\item{report}{the report object}

\item{to_numeric}{a vector of column names}

\item{scan_res}{a data frame obtained from the \code{scan_data()} function}
}
\value{
the input data frame where all the specified or the detected columns
(as described above) have been converted into numeric
}
\description{
When the names of the columns to be converted is not provided, the function
looks for columns that contains only missing, numeric and character values
and where the percentage of numeric values is at least 2 times the percent of
character values.
}
\examples{
data = readRDS(system.file("extdata", "messy_data.RDS",
               package = "cleanepi"))
to_numeric = "age"
}
