% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardisation_helpers.R
\name{date_check_column_existence}
\alias{date_check_column_existence}
\title{Check if date column exists in the given dataset}
\usage{
date_check_column_existence(data, date_column_names)
}
\arguments{
\item{data}{A dataframe}

\item{date_column_names}{The name of the columns to check}
}
\description{
Check if date column exists in the given dataset
}
\keyword{internal}
