% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardization_helpers.R
\name{date_check_timeframe}
\alias{date_check_timeframe}
\title{Check date time frame}
\usage{
date_check_timeframe(first_date, last_date)
}
\arguments{
\item{first_date}{A Date object specify.ing the first valid date.
The default value is fifty years before the \code{last_date}}

\item{last_date}{A  Date object specifying the last valid date.
The defaults value is the current date.}
}
\value{
A list with the first and last dates
}
\description{
Check date time frame
}
\keyword{internal}
