% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardization_helpers.R
\name{date_check_timeframe}
\alias{date_check_timeframe}
\title{Check date time frame}
\usage{
date_check_timeframe(first_date, last_date)
}
\arguments{
\item{first_date}{A Date object specifying the first valid date.
The arbitrary default value is fifty years before the \code{last_date}.
This can also be a character in ISO8601 format i.e. "2024-12-31".}

\item{last_date}{A Date object specifying the last valid date.
The defaults value is the current date. This can also be a character in
ISO8601 format i.e. "2024-12-31".}
}
\value{
A list with the first and last dates
}
\description{
Check date time frame
}
\keyword{internal}
