% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardisation_helpers.R
\name{date_convert}
\alias{date_convert}
\title{Convert characters to dates}
\usage{
date_convert(data, cols, error_tolerance, format = NULL, timeframe = NULL)
}
\arguments{
\item{data}{the input data frame}

\item{cols}{the date column name(s)}

\item{error_tolerance}{a number between 0 and 1 indicating the proportion of
entries which cannot be identified as dates to be tolerated. See the
\code{clean_data()} helper for more details.}

\item{format}{the format of the date values in the date columns}

\item{timeframe}{a vector of 2 values of type date. If provided, date values
that do not fall within this timeframe will be set to \code{NA}.}
}
\value{
the input data frame where the specified columns have been converted
into Date.
}
\description{
Convert characters to dates
}
\keyword{internal}
