% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardization_helpers.R
\name{date_detect_day_or_month}
\alias{date_detect_day_or_month}
\title{Detect whether it's day or month}
\usage{
date_detect_day_or_month(x)
}
\arguments{
\item{x}{A string of interest}
}
\description{
Detect whether it's day or month
}
\keyword{internal}
