% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardization_helpers.R
\name{date_detect_format}
\alias{date_detect_format}
\title{Detect a date format with only 1 separator}
\usage{
date_detect_format(x)
}
\arguments{
\item{x}{A \code{<character>} with the string of interest}
}
\value{
A \code{<character>} with the identified format.
}
\description{
Detect a date format with only 1 separator
}
\keyword{internal}
