% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardisation_helpers.R
\name{date_detect_separator}
\alias{date_detect_separator}
\title{Detect the special character that is the separator in the date values}
\usage{
date_detect_separator(x)
}
\arguments{
\item{x}{A string of interest}
}
\value{
A detected separator
}
\description{
Detect the special character that is the separator in the date values
}
\keyword{internal}
