% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardization_helpers.R
\name{date_get_format}
\alias{date_get_format}
\title{Infer date format from a vector or characters}
\usage{
date_get_format(x)
}
\arguments{
\item{x}{The input character vector}
}
\value{
A string with the inferred date format
}
\description{
Infer date format from a vector or characters
}
\keyword{internal}
