% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_get_part1}
\alias{date_get_part1}
\title{Split a string based on a pattern and return the first element of the
resulting vector.}
\usage{
date_get_part1(x, sep)
}
\arguments{
\item{x}{A \code{<character>} or \code{<Date>} value}

\item{sep}{A \code{<character>} with the separator identified in 'x'}
}
\value{
A \code{<character>} with the first element of the vector returned
by the \code{strsplit()} function.
}
\description{
Split a string based on a pattern and return the first element of the
resulting vector.
}
\keyword{internal}
