% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_get_part2}
\alias{date_get_part2}
\title{Get part2 of date value}
\usage{
date_get_part2(x, sep)
}
\arguments{
\item{x}{A string of interest}

\item{sep}{A separator in the string of interest}
}
\description{
Get part2 of date value
}
\keyword{internal}
