% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardisation_helpers.R
\name{date_guess_convert}
\alias{date_guess_convert}
\title{Guess if a character vector contains Date values, and convert them to date}
\usage{
date_guess_convert(data, error_tolerance, timeframe)
}
\arguments{
\item{data}{A data frame}

\item{error_tolerance}{A numeric value between 0 and 1 that signifies the
proportion of entries that cannot be recognized as dates and are acceptable.
For more detailed information, you can refer to the \code{clean_data()}
helper function.}

\item{timeframe}{The expected first and last date. See the \code{clean_data()}
helper for more details.}
}
\description{
Guess if a character vector contains Date values, and convert them to date
}
\keyword{internal}
