% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardisation_helpers.R
\name{date_make_format}
\alias{date_make_format}
\title{Build the auto-detected format}
\usage{
date_make_format(f1, f2, tmp_sep)
}
\arguments{
\item{f1}{The first part of the date values}

\item{f2}{The second part of the date values}

\item{tmp_sep}{The character string that separate the first and second parts
of the date values.}
}
\value{
A character string that represent the inferred format of the date
values.
}
\description{
Put together the different date format characters that were identified in
the target date column.
}
\keyword{internal}
