% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardization_helpers.R
\name{date_make_format}
\alias{date_make_format}
\title{Build the auto-detected format}
\usage{
date_make_format(f1, f2, f3)
}
\arguments{
\item{f1}{The first part of the inferred format}

\item{f2}{The second part of the inferred format}

\item{f3}{The third part of the inferred format}
}
\value{
A character string that represent the inferred format from the date
values. It returns NULL when the format was not resolved.
}
\description{
Put together the different date format characters that were identified from
the target date column.
}
\keyword{internal}
