% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardisation_helpers.R
\name{date_process}
\alias{date_process}
\title{Process date variable}
\usage{
date_process(x)
}
\arguments{
\item{x}{an object of date class}
}
\value{
the modified input object
}
\description{
Process date variable
}
\keyword{internal}
