% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardization_helpers.R
\name{date_process}
\alias{date_process}
\title{Process date variable}
\usage{
date_process(x)
}
\arguments{
\item{x}{A \code{<Date>} object}
}
\value{
The converted input value into \code{<Date>} or \code{<character>}
}
\description{
Process date variable
}
\keyword{internal}
