% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_dates.R
\name{date_rescue_lubridate_failures}
\alias{date_rescue_lubridate_failures}
\title{Find the dates that lubridate couldn't}
\usage{
date_rescue_lubridate_failures(date_a_frame, original_dates, column_name)
}
\arguments{
\item{date_a_frame}{A \code{<data.frame>} where each column contains a
different parsing of the same date vector}

\item{original_dates}{A \code{<vector>} of original dates}

\item{column_name}{A \code{<character>} with the target column name}
}
\value{
A \code{<list>} with the following two elements: the input data
frame where the values that do not match the proposed formats have been
converted into Date, and a boolean that informs about the presence of
ambiguous values or not.
}
\description{
Find the dates that lubridate couldn't
}
\keyword{internal}
