% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardisation_helpers.R
\name{date_trim_outliers}
\alias{date_trim_outliers}
\title{Trim dates outside of the defined boundaries}
\usage{
date_trim_outliers(new_dates, dmin, dmax, cols, original_dates)
}
\arguments{
\item{new_dates}{a vector of the new date values}

\item{dmin}{the minimum date}

\item{dmax}{the maximum date}

\item{cols}{the name of the date column of interest}

\item{original_dates}{a vector of the original date values}
}
\description{
Trim dates outside of the defined boundaries
}
\keyword{internal}
