% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_cleanepi.R
\name{detect_date_format}
\alias{detect_date_format}
\title{function to detect the date format with only 1 separator}
\usage{
detect_date_format(x)
}
\description{
function to detect the date format with only 1 separator
}
