% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_standardisation_helpers.R
\name{detect_date_separator}
\alias{detect_date_separator}
\title{Detect the special character that is the separator in the date values}
\usage{
detect_date_separator(x)
}
\arguments{
\item{x}{the string of interest}
}
\value{
the detected separator
}
\description{
Detect the special character that is the separator in the date values
}
