% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary_based_cleaning.R
\name{dictionary_make_metadata}
\alias{dictionary_make_metadata}
\title{Make data dictionary for 1 field}
\usage{
dictionary_make_metadata(x, field_column)
}
\arguments{
\item{x}{A \code{<vector>} of choices from the input data}

\item{field_column}{A \code{<character>} with the name of the field where the
choices are made.}
}
\value{
A \code{<data.frame>} with the dictionary in the format that is
accepted by the \pkg{matchmaker} package.
}
\description{
Make data dictionary for 1 field
}
\keyword{internal}
