% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/find_and_remove_duplicates.R,
%   R/remove_duplicates.R
=======
% Please edit documentation in R/find_and_remove_duplicates.R
>>>>>>> 7ed29d1 (add scan_data(), and check_ids_uniqueness() functions)
\name{find_duplicates}
\alias{find_duplicates}
\title{Identify and return duplicated rows in a data frame or linelist.}
\usage{
find_duplicates(data, target_columns)

find_duplicates(data, target_columns)
}
\arguments{
\item{data}{The input data frame or linelist.}

\item{target_columns}{A vector of columns names or indices to consider when
looking for duplicates. When the input data is a \code{linelist} object,
this parameter can be set to \code{tags} if you wish to look for
duplicates across the tagged variables only.}
}
\value{
Data frame or linelist of all duplicated rows with following 2
additional columns:
\enumerate{
\item \code{row_id}: the indices of the duplicated rows from the input data.
Users can choose from these indices, which row they consider as
redundant in each group of duplicates.
\item \code{group_id}: a unique identifier associated to each group of
duplicates.
}

Data frame or linelist of all duplicated rows with following 2
additional columns:
\enumerate{
\item \code{row_id}: the indices of the duplicated rows from the input data.
Users can choose from these indices, which row they consider as
redundant in each group of duplicates.
\item \code{group_id}: a unique identifier associated to each group of
duplicates.
}
}
\description{
Identify and return duplicated rows in a data frame or linelist.

Identify and return duplicated rows in a data frame or linelist.
}
\examples{
dups <- find_duplicates(
  data = readRDS(system.file("extdata", "test_linelist.RDS",
    package = "cleanepi")),
<<<<<<< HEAD
  target_columns = c("dt_onset", "dt_report", "sex", "outcome")
)

dups <- find_duplicates(
  data = readRDS(system.file("extdata", "test_linelist.RDS",
    package = "cleanepi"
  )),
=======
>>>>>>> f1c90a5 (add the convert_to_numeric function and restructure the package)
  target_columns = c("dt_onset", "dt_report", "sex", "outcome")
)

}
