% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_cleanepi_settings.R
\name{get_default_params}
\alias{get_default_params}
\title{Set and return \code{clean_data()} default parameters}
\usage{
get_default_params()
}
\value{
The list of the default cleaning parameters.
}
\description{
When \code{clean_data()} function is called without any argument, these
default values provided to the function's arguments will be applied on the
input data. By default, operations that require the target columns to be
specified by the user will not be performed. The default cleaning operations
include: i) standardizing column names, ii) detecting and removing
duplicates, and iii) removing constant data.
}
\keyword{internal}
