% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_and_remove_duplicates.R
\name{get_target_column_names}
\alias{get_target_column_names}
\title{Get the names of the columns from which duplicates will be found}
\usage{
get_target_column_names(data, target_columns, cols)
}
\arguments{
\item{data}{A dataframe or linelist}

\item{target_columns}{A vector of column names}

\item{cols}{A vector of empty and constant columns}
}
\value{
A vector with the target column names or indexes
}
\description{
Get the names of the columns from which duplicates will be found
}
\keyword{internal}
