% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{get_target_column_names}
\alias{get_target_column_names}
\title{Get the names of the columns from which duplicates will be found}
\usage{
get_target_column_names(target_columns, data)
}
\arguments{
\item{target_columns}{the user specified target column name}

\item{data}{the input dataset}
}
\value{
a \code{vector} with the target column names or indexes
}
\description{
Get the names of the columns from which duplicates will be found
}
