% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_date_sequence.R
\name{is_date_sequence_ordered}
\alias{is_date_sequence_ordered}
\title{Check order of a sequence of date-events}
\usage{
is_date_sequence_ordered(x)
}
\arguments{
\item{x}{A \code{<vector>} of \code{<Date>} values}
}
\value{
\code{TRUE} if elements of the vector are ordered, \code{FALSE} otherwise.
}
\description{
Check order of a sequence of date-events
}
\keyword{internal}
