% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary_based_cleaning.R
\name{make_readcap_dictionary}
\alias{make_readcap_dictionary}
\title{Convert Redcap data dictionary into {matchmaker} dictionary format}
\usage{
make_readcap_dictionary(metadata, field_column, opt_column, field_type)
}
\arguments{
\item{metadata}{A data frame with the data dictionary associated to a
Redcap project}

\item{field_column}{The name of the column, in the input dictionary, that
contains the field names in of the Redcap project data}

\item{opt_column}{The name of the column, in the input dictionary, that
contains the definition of the choices in every column of the Redcap
project data}

\item{field_type}{A \code{character} with the name of the column that contains the
field type information}
}
\value{
a data frame with 4 columns. This is in the format required by the
{matchmaker} R package for dictionary-based cleaning.
}
\description{
Convert Redcap data dictionary into {matchmaker} dictionary format
}
\examples{
test <- make_readcap_dictionary(
  metadata     = system.file("extdata", "test_readcap_dictionary.RDS",
                             package = "cleanepi"),
  field_column = "field_name",
  opt_column   = "select_choices_or_calculations",
  field_type   = "field_type"
)


}
\keyword{internal}
