% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numbers_only}
\alias{numbers_only}
\title{Detects whether a string contains only numbers or not.}
\usage{
numbers_only(x)
}
\arguments{
\item{x}{A \code{<character>} value}
}
\value{
\code{TRUE} if the string only contains numbers, \code{FALSE}
otherwise
}
\description{
Detects whether a string contains only numbers or not.
}
\keyword{internal}
