% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numbers_only}
\alias{numbers_only}
\title{Detect whether a string contains only numbers}
\usage{
numbers_only(x)
}
\arguments{
\item{x}{the string that contains the numbers}
}
\description{
Detect whether a string contains only numbers
}
