% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary_based_cleaning.R
\name{print_misspelled_values}
\alias{print_misspelled_values}
\title{Print the detected misspelled values}
\usage{
print_misspelled_values(misspelled_options)
}
\arguments{
\item{misspelled_options}{A list with the misspelled values found in
the different columns of the input data.}
}
\description{
Print the detected misspelled values
}
\keyword{internal}
