% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_report.R
\name{print_report}
\alias{print_report}
\title{Generate report from data cleaning operations}
\usage{
print_report(
  data,
  report_title = "{cleanepi} data cleaning report",
  output_directory = ".",
  output_filename = NULL,
  format = "html",
  print = TRUE
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{linelist} object returned from the
\code{\link[=clean_data]{clean_data()}} or the main functions of each data cleaning module.}

\item{report_title}{The title to appear on the report}

\item{output_directory}{A string specifying directory in which to save the
report. Default is current directory.}

\item{output_filename}{A string specifying filename for the report, excluding
any file extension. If no filename is supplied, one will be automatically
generated with the format \code{cleanepi_report_YYMMDD_HHMMSS}.}

\item{format}{The file format of the report. Currently only \code{"html"}
is supported.}

\item{print}{A logical that specifies whether to print the generated HTML
file or no.}
}
\value{
A string containing the name and path of the saved report
}
\description{
Generate report from data cleaning operations
}
\examples{
\donttest{
replace_missing_values = list(target_columns = NULL, na_strings = "-99")

remove_duplicates = list(target_columns   = NULL,
                         rm_empty_rows    = TRUE,
                         rm_empty_cols    = TRUE,
                         rm_constant_cols = TRUE)

standardize_date = list(target_columns  = NULL,
                        error_tolerance = 0.5,
                        format          = NULL,
                        timeframe       = as.Date(c("1973-05-29",
                                                    "2023-05-29")))

standardize_subject_ids = list(id_col_name = "study_id",
                               format      = NULL,
                               prefix      = "PS",
                               suffix      = "P2",
                               range       = c(1, 100))

# scan through the data
scan_res <- scan_data(data = readRDS(system.file("extdata", "test_df.RDS",
                                                 package = "cleanepi")))

# Perform data cleaning
cleaned_data <- clean_data(
  data   = readRDS(system.file("extdata", "test_df.RDS",
                               package = "cleanepi")),
  params = list(
    keep                    = NULL,
    replace_missing_values  = replace_missing_values,
    remove_duplicates       = remove_duplicates,
    standardize_date        = standardize_date,
    standardize_subject_ids = standardize_subject_ids,
    to_numeric              = "sex",
    dictionary              = NULL
  )
)

# add the data scanning result to the report
cleaned_data <- add_to_report(x     = cleaned_data,
                              key   = "scanning_result",
                              value = scan_res)

# save a report in the current directory using the previously-created objects
print_report(
  data             = cleaned_data,
  report_title     = "{cleanepi} data cleaning report",
  output_directory = ".",
  output_filename  = "template_data_report",
  format           = "html",
  print            = TRUE
)
\dontshow{file.remove("./template_data_report.html")}
}

}
