% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_report.R
\name{print_report}
\alias{print_report}
\title{Print the report from the data cleaning operations}
\usage{
print_report(report)
}
\arguments{
\item{report}{the report object}
}
\description{
Print the report from the data cleaning operations
}
