% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/find_and_remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicates based on selected columns from a data frame
or linelist object.}
\usage{
remove_duplicates(data, target_columns, remove = NULL, report = list())
}
\arguments{
\item{data}{The input data frame.}

\item{target_columns}{A vector of column names to use when looking for
duplicates. When the input data is a \code{linelist} object, this
parameter can be set to \code{tags} if you wish to look for duplicates on
tagged columns.}

\item{remove}{A vector of duplicate indices to be removed.
Duplicate indices are unique identifiers for all rows in the original
data frame or linelist that are duplicates of each other based on the
\code{target_columns}.
If remove = NULL (default value), the first duplicate is kept and
the rest of the duplicates in the group are removed.}

\item{report}{A list with the information about the effects of the
cleaning steps.}
}
\value{
A \code{list} with elements data (the filtered dataset) and report.
}
\description{
Remove duplicates based on selected columns from a data frame
or linelist object.
}
\examples{
no_dups <- remove_duplicates(
  data           = readRDS(system.file("extdata", "test_linelist.RDS",
                             package = "cleanepi")),
  target_columns = "tags",
  remove         = NULL,
  report         = list()
=======
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicates based on selected columns from a data frame
or linelist object.}
\usage{
remove_duplicates(data, target_columns, remove = NULL, report = list())
}
\arguments{
\item{data}{The input data frame.}

\item{target_columns}{A vector of column names to use when looking for
duplicates. When the input data is a \code{linelist} object, this
parameter can be set to \code{tags} if you wish to look for duplicates on
tagged columns.}

\item{remove}{A vector of duplicate indices to be removed.
Duplicate indices are unique identifiers for all rows in the original
data frame or linelist that are duplicates of each other based on the
\code{target_columns}.
If remove = NULL (default value), the first duplicate is kept and
the rest of the duplicates in the group are removed.}

\item{report}{A list with the information about the effects of the
cleaning steps.}
}
\value{
A \code{list} with elements data (the filtered dataset) and report.
}
\description{
Remove duplicates based on selected columns from a data frame
or linelist object.
}
\examples{
no_dups <- remove_duplicates(
  data = readRDS(system.file("extdata", "test_linelist.RDS",
    package = "cleanepi"
  )),
  target_columns = "tags",
  report = list()
>>>>>>> 433caea (update NAMESPACE after devtools::document())
)

}
