% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicates from a data frame or linelist object}
\usage{
remove_duplicates(data, duplicates_from, remove = -1, report = list())
}
\arguments{
\item{data}{the input data frame}

\item{duplicates_from}{a vector of columns names to use when looking for
duplicates. When the input data is a \code{linelist} object, this
parameter can be set to \code{tags} if you wish to look for duplicates on
tagged variables. Only used when \code{remove_duplicates=TRUE}}

\item{remove}{a vector of the indices of the duplicated rows to be removed.
Default is \code{-1} i.e. to keep only the first instance of duplicated rows.}

\item{report}{a list with the information about the effects of the
cleaning steps}
}
\value{
a list with 2 elements: the filtered dataset and the report object
}
\description{
Remove duplicates from a data frame or linelist object
}
\examples{
no_dups <- remove_duplicates(
  data = readRDS(system.file("extdata", "test_linelist.RDS",
  package = "cleanepi")),
  duplicates_from = "tags",
  remove = -1,
  report = list()
)

}
