% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_missing_values.R
\name{replace_missing_values}
\alias{replace_missing_values}
\title{Replace missing values with \code{NA}}
\usage{
replace_missing_values(data, from = NULL, na_comes_as = NULL)
}
\arguments{
\item{data}{the input data}

\item{from}{a vector of column names. If provided, the missing values
substitution will only be performed in those columns}

\item{na_comes_as}{the string that represents the missing values in the
column of interest}
}
\value{
the input data where missing values are replaced by \code{NA}
}
\description{
Replace missing values with \code{NA}
}
\examples{
cleaned_data <- replace_missing_values(
  data        = readRDS(system.file("extdata", "test_df.RDS",
                                    package = "cleanepi")),
  from        = "sex",
  na_comes_as = "-99"
)

}
