% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_missing_values.R
\name{replace_missing_values}
\alias{replace_missing_values}
\title{Replace missing values with \code{NA}}
\usage{
replace_missing_values(
  data,
  target_columns = NULL,
  na_strings = cleanepi::common_na_strings
)
}
\arguments{
\item{data}{The input \code{<data.frame>} or \code{<linelist>}}

\item{target_columns}{A \code{<vector>} of column names. If provided, missing
values will be substituted only in the specified columns. When the input
data is a \code{<linelist>} object, this parameter can be set to
\code{linelist_tags} to replace missing values with \code{NA} in the
tagged columns only.}

\item{na_strings}{A \code{<vector>} of characters that represent the missing
values in the columns of interest. By default, it utilizes
\code{cleanepi::common_na_strings}. However, if the missing values string
in the columns of interest is not included in this predefined vector,
it can be used as the value for this argument. Matching of \code{data} with
\code{na_strings} is insensitive to case and whitespace.}
}
\value{
The input data where missing values are replaced by \code{NA}.
}
\description{
Replace missing values with \code{NA}
}
\examples{
data <- readRDS(
  system.file("extdata", "test_df.RDS", package = "cleanepi")
)

# replace all occurrences of '-99' with NA
cleaned_data <- replace_missing_values(
  data = data,
  target_columns = NULL,
  na_strings = "-99"
)

# print the names of the columns where the replacement occurred
print_report(cleaned_data, "missing_values_replaced_at")

}
