% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_name_standardisation.R
\name{standardize_column_names}
\alias{standardize_column_names}
\title{Standardize column names of a data frame or linelist}
\usage{
standardize_column_names(x, keep = NULL)
}
\arguments{
\item{x}{A data frame or linelist}

\item{keep}{A vector of column names to keep as they are. The Default
is \code{NULL}.}
}
\value{
A data frame with knit column names
}
\description{
Standardize column names of a data frame or linelist
}
\examples{
cleaned_data <- standardize_column_names(
  x    = readRDS(system.file("extdata", "test_df.RDS",
                             package = "cleanepi")),
  keep = "dateOfBirth"
)

}
