% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_date.R
\name{standardize_dates}
\alias{standardize_dates}
\title{Standardize date variables}
\usage{
standardize_dates(
  data,
  target_columns = NULL,
  format = NULL,
  timeframe = NULL,
  error_tolerance = 0.5
)
}
\arguments{
\item{data}{A data frame or linelist}

\item{target_columns}{A vector or a comma-separated list of the target date
column names. When the input data is a \code{linelist} object, this parameter
can be set to \code{linelist_tags} if you wish to standardize the date columns
across tagged columns only.}

\item{format}{A format of the date values in the date columns}

\item{timeframe}{A vector of 2 values of type date. If provided, date values
that do not fall within this timeframe will be set to \code{NA}.}

\item{error_tolerance}{A numerical value between 0 and 1, indicating the
proportion of entries which cannot be identified as dates to be tolerated.
See the \code{clean_data()} helper for more details.}
}
\value{
The input dataset where the date columns have been standardized
}
\description{
Standardize date variables
}
\examples{
dat <- standardize_dates(
  data            = readRDS(system.file("extdata", "test_df.RDS",
                                        package = "cleanepi")),
  target_columns  = "date_first_pcr_positive_test",
  format          = NULL,
  timeframe       = NULL,
  error_tolerance = 0.5
)
}
