% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-potools.R
\name{tr_}
\alias{tr_}
\title{Flag out what message will be translated using \{potools\}}
\usage{
tr_(...)
}
\arguments{
\item{...}{A character string. This represents the message to be translated}
}
\value{
The input object
}
\description{
Flag out what message will be translated using \{potools\}
}
\details{
This function was taken from the \verb{Translation for package developers}
vignette of the \code{\{potools\}} package.
}
\keyword{internal}
